+++ src/server/routers/index.py
@@ -3,7 +3,6 @@ import logging
 
 from src.server.routers.figure_and_relation import figure_and_relation_router
 from src.server.routers.fine_grained_feed import fine_grained_feed_router
-from src.server.routers.graph import graph_router
 from src.server.routers.knowledge import knowledge_router
 from src.server.routers.user import user_router
 
@@ -23,6 +22,5 @@ async def registerRouters(app: Robyn):
     app.include_router(figure_and_relation_router)
     app.include_router(fine_grained_feed_router)
     app.include_router(knowledge_router)
-    app.include_router(graph_router)
 
     app.get("/ping")(lambda: "pong")